+++ src/test/java/ch/epfl/leb/alica/acpack/analyzers/integrator/IntegratorTest.java
@@ -17,7 +17,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package ch.epfl.leb.alica.analyzers.integrator;
+package ch.epfl.leb.alica.acpack.analyzers.integrator;
 
 import ch.epfl.leb.alica.acpack.analyzers.integrator.Integrator;
 import org.junit.Test;